Add support for the Hermes JS engine.
  This allows users to toggle between Duktape and Hermes as the
  underlying JS engine/interpreter used by React-JUCE. Hermes provides a
  significant performance advantage over Duktape but it's inclusion is
  somewhat experimental and has a little more nuance in terms of it's
  build environment.

  See: https://github.com/facebook/hermes
